# Replace `indexOf === 0` with `startsWith`