A small PR.

It is more readable to call `startsWith` than to call `indexOf` and compare the result with zero to determine whether a string starts with a given prefix.

`startsWith` is also more efficient as it only compares at the beginning of the string, while `indexOf` searches the entire string.