@ai 2024-08-06
Can you run https://github.com/postcss/benchmark before and after the change?